Added coast mode to the turret after 5 seconds
-Also changed the shotoerdata to an int instead of a double

+++ src/main/java/frc/robot/commands/Shooter/FiringCommand.java
@@ -116,8 +116,8 @@ public class FiringCommand extends CommandBase {
         double distanceToTarget = m_limelight.getDistanceToTarget();
         int closestPoint = InterpolatedData.closestPoint(distanceToTarget, kShooter.kShooterData.shooterSteps);
 
-        double dataX[] = kShooter.kShooterData.shooterDataX;
-        double dataY[] = kShooter.kShooterData.shooterDataY;
+        int dataX[] = kShooter.kShooterData.shooterDataX;
+        int dataY[] = kShooter.kShooterData.shooterDataY;
 
         double interpolatedSpeed = InterpolatedData.getInterpolatedSpeed(
             dataX[closestPoint],
